pyproject.toml
[tool.poetry]
name = "adaptive-rag-chatbot"
version = "0.1.0"
description = ""
authors = ["Steve Cohen"]
readme = "README.md"
packages = [{ include = "graph" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
beautifulsoup4 = "^4.12.3"
chromadb = "^0.5.0"
python-dotenv = "^1.0.1"
black = "^24.4.2"
isort = "^5.13.2"
pytest = "^8.2.0"
langgraph-cli = "^0.1.52"
tiktoken = "^0.7.0"
langchain-openai = "^0.1.12"
pypdf = "^4.3.1"
faiss-cpu = "^1.8.0.post1"
langchain = "^0.2.12"
langchain-community = "^0.2.11"
langchainhub = "^0.1.20"
langgraph = "^0.2.3"
langgraph-checkpoint-sqlite = "^1.0.0"
tavily-python = "^0.4.0"
langchain-chroma = "^0.1.3"
langchain-pinecone = "^0.1.3"
pinecone-client = "^5.0.1"
streamlit = "^1.38.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
